backend/HackGSUBhojan/BhojanTwilio/requirements.txt
azure-functions==1.0.4
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
firebase-admin==3.1.0
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.5.0
google-cloud-storage==1.21.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.3
httplib2==0.14.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
msgpack==0.6.2
numpy==1.17.3
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyJWT==1.7.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.12.0
twilio==6.33.0
uritemplate==3.0.0
urllib3==1.25.6
virtualenv==16.7.7
